# develop/api/swagger-spec.json
{"swagger":"2.0","info":{"description":"The API description.","version":"1.0","title":"App example"},"basePath":"/","tags":[{"name":"app","description":""}],"schemes":["http"],"securityDefinitions":{"bearer":{"type":"apiKey","name":"Authorization","in":"header"}},"paths":{"/":{"get":{"responses":{"200":{"description":""}},"produces":["application/json"],"consumes":["application/json"]}},"/api/auth/authenticate":{"post":{"responses":{"201":{"description":""}},"tags":["/api/auth"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/api/auth/data":{"get":{"responses":{"200":{"description":""}},"tags":["/api/auth"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/api/users":{"post":{"parameters":[{"name":"CreateUserDto","required":true,"in":"body","schema":{"$ref":"#/definitions/CreateUserDto"}}],"responses":{"201":{"description":""}},"tags":["/api/users"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"get":{"responses":{"200":{"description":""}},"tags":["/api/users"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/api/users/protected":{"get":{"responses":{"200":{"description":""}},"tags":["/api/users"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/api/users/{id}":{"get":{"parameters":[{"type":"string","name":"id","required":true,"in":"path"}],"responses":{"200":{"description":""}},"tags":["/api/users"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/api/cats":{"post":{"parameters":[{"name":"CreateCatDto","required":true,"in":"body","schema":{"$ref":"#/definitions/CreateCatDto"}}],"responses":{"201":{"description":""}},"tags":["/api/cats"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"get":{"responses":{"200":{"description":""}},"tags":["/api/cats"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/api/cats/protected":{"get":{"responses":{"200":{"description":""}},"tags":["/api/cats"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/api/cats/{id}":{"get":{"parameters":[{"type":"string","name":"id","required":true,"in":"path"}],"responses":{"200":{"description":""}},"tags":["/api/cats"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}}},"definitions":{"CreateUserDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"number"},"role":{"type":"number"}},"required":["username","password","role"]},"CreateCatDto":{"type":"object","properties":{"name":{"type":"string"},"age":{"type":"number"},"breed":{"type":"string"}},"required":["name","age","breed"]}}}